requirements.txt
streamlit==1.37.1
python-dotenv==1.0.1
pydantic==2.8.2
pandas==2.2.2
requests==2.32.3

#for file parsing
pypdf==4.3.1
python-docx==1.1.2
chardet==5.2.0
pytesseract
pdf2image
Pillow
